feat: Add the new abilities and upgrades from WD 475
This WD introduced a new campaign system with a new set of ace abilities and
aircraft upgrades. While the abilities are supposed to be used in the campaign,
the article does mention using them in one off games. Added the new ace
abilities into the existing Expanded Ace Ability list, and created a new list
for the aircraft upgrades.

The aircraft upgrades do not do any automatic modification of the stat lines of
the planes, nor do they do any complicated error checking (e.g. checking if the
max speed is already 9 when the Upgraded Engines are purchased).

Also fixed a bug that required you to select the Heavy Death Ray for the Doom
Scythe, even though it's automatically included.